+++ packages/core/src/browser/shell/tab-bars.ts
@@ -305,6 +305,7 @@ export class ToolbarAwareTabBar extends ScrollableTabBar {
 
         super(options);
         this.rewireDOM();
+        this.tabBarToolbarRegistry.onDidChange(() => this.update());
     }
 
     /**